tests: Move GPU-AV test to dedicated files
