+++ tests/CMakeLists.txt
@@ -85,6 +85,10 @@ target_sources(vk_layer_validation_tests PRIVATE
     unit/geometry_tessellation.cpp
     unit/geometry_tessellation_positive.cpp
     unit/gpu_av.cpp
+    unit/gpu_av_buffer_device_address.cpp
+    unit/gpu_av_descriptor_indexing.cpp
+    unit/gpu_av_indirect_buffer.cpp
+    unit/gpu_av_oob.cpp
     unit/gpu_av_positive.cpp
     unit/graphics_library.cpp
     unit/graphics_library_positive.cpp
